chore: Rename uuidprovider -> accountprovider
